+++ changelog.md
@@ -11,6 +11,7 @@ vX.X.X
 - Added ability to select between "overwrite" and "append" logging methods, closes #99.
 - Added option for user to choose termination character(s), closes #82.
 - Added support for ANSI escape sequences (in particular, the colour codes), closes #100.
+- Fixed bug where buffer limit was not being obeyed for the RX data when stored in ObservableList of Nodes, closes #101.
 
 v0.4.1
 ------